+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- feat: hide cursor if editor don't have the focus.
+
 ## [0.3.1] - 2026-01-11
 
 ### Fixed